pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bharat-highway-asset-intelligence"
version = "0.1.0"
description = "Research-first, official-first Bharat Highway intelligence scaffold"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "pyyaml>=6.0",
  "requests>=2.32",
  "python-dateutil>=2.9",
  "pydantic>=2.8",
  "pyarrow>=14",
  "duckdb>=1.1"
]
